# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bid Arena API
  version: 0.1.0
  description: |
    拍卖间 HTTP 契约。金额为整数积分，ID 使用字符串，服务端是唯一裁判。

    所有响应（含 4xx/5xx）都是 Envelope 封套。出价的幂等键以下面 BidRequest.requestId
    为准；Idempotency-Key 头是等价的可选写法，两者都提供时必须一致。

    两个约定，写在前端接入前：
    1) 【字段缺失】与【值为 null】是同一件事：客户端不得用 in/keys 判断字段是否存在。
       服务端**可能**输出 `"winner": null`，也**可能**直接省略它（例：取消的拍卖结果里没有 winner），
       两者的语义都是“这一场没有中标人”。不要依赖其中之一，只判断“取不到或为 null”。
    2) 成功码有两种：OK 与 IDEMPOTENCY_REPLAY（都是 HTTP 200）。重放不是错误，
       但客户端应当用它提示“本次提交已按首次结果处理”，而不是当成新成交。

    实时事件走 WebSocket，不在本文件里描述消息体：事件信封与事件类型见
    docs/REALTIME_AND_COMMAND_FLOW.md（§3 信封、§4 事件表、§6 缺口恢复）。
    这里只固定它的**入口**：POST /auth/ws-tickets 换一次性票，再用票连 WS。
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /users/me:
    get:
      tags: [Auth]
      security: [{bearerAuth: []}]
      operationId: currentUser
      responses:
        '200': {$ref: '#/components/responses/EnvelopeUser'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /auctions:
    get:
      tags: [Auctions]
      security: [{bearerAuth: []}]
      operationId: listAuctions
      parameters:
        - in: query
          name: status
          schema: {$ref: '#/components/schemas/AuctionStatus'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200': {$ref: '#/components/responses/EnvelopeAuctionPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auctions/{auctionId}:
    get:
      tags: [Auctions]
      security: [{bearerAuth: []}]
      operationId: getAuction
      parameters: [{$ref: '#/components/parameters/AuctionId'}]
      responses:
        '200': {$ref: '#/components/responses/EnvelopeAuction'}
        '404': {$ref: '#/components/responses/NotFound'}
  /auctions/{auctionId}/join:
    post:
      tags: [Auctions]
      security: [{bearerAuth: []}]
      operationId: joinAuction
      parameters: [{$ref: '#/components/parameters/AuctionId'}]
      responses:
        '200': {$ref: '#/components/responses/EnvelopeParticipant'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /auctions/{auctionId}/bids:
    get:
      tags: [Bids]
      security: [{bearerAuth: []}]
      operationId: listBids
      parameters:
        - {$ref: '#/components/parameters/AuctionId'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200': {$ref: '#/components/responses/EnvelopeBidPage'}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      tags: [Bids]
      security: [{bearerAuth: []}]
      operationId: placeBid
      description: |
        提交出价。幂等键以请求体的 requestId 为准；也可以用 Idempotency-Key 头代替。
        重复提交同一个幂等键时返回**首次**结果，code 为 IDEMPOTENCY_REPLAY（HTTP 仍是 200），
        客户端据此可以安全重试。
      parameters:
        - {$ref: '#/components/parameters/AuctionId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BidRequest'}
      responses:
        '200': {$ref: '#/components/responses/EnvelopeBidResult'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /auctions/{auctionId}/result:
    get:
      tags: [Auctions]
      security: [{bearerAuth: []}]
      operationId: auctionResult
      parameters: [{$ref: '#/components/parameters/AuctionId'}]
      responses:
        '200': {$ref: '#/components/responses/EnvelopeResult'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/auctions:
    post:
      tags: [Admin]
      security: [{bearerAuth: []}]
      operationId: createAuction
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAuctionRequest'}
      responses:
        '201': {$ref: '#/components/responses/EnvelopeAuction'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /agent/auctions/{auctionId}:
    servers:
      - url: http://localhost:8090/api/v1
    get:
      tags: [Agent]
      security: [{agentToken: []}]
      operationId: agentAuction
      description: |
        读取竞拍快照。**这把接口同时也在 8080 上可用**；8090 只是额外的一个最小暴露面，
        它上面除了 /agent/** 之外的一切路径都返回 404 封套（连 /health 也没有）。

        Agent 凭证的三个失败码分工：401 = 不认识/已过期/已吊销（不区分原因，避免试探）；
        403 = 凭证有效但越权（缺权限，或这场拍卖不在 auctionIds 里）；429 = 超出该 Token 的频率上限。
      parameters: [{$ref: '#/components/parameters/AuctionId'}]
      responses:
        '200': {$ref: '#/components/responses/EnvelopeAuction'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    agentToken: {type: http, scheme: bearer, bearerFormat: AUCTION_AGENT_TOKEN}
  parameters:
    AuctionId: {in: path, name: auctionId, required: true, schema: {type: string}}
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: |
        幂等键的可选写法。优先使用 BidRequest.requestId；只有本头而无 requestId 时以本头为准，
        两者同时出现且不一致时返回 400 VALIDATION_FAILED（因为无法判断调用方真实意图，
        猜错会把一次重试当成新出价，造成重复冻结）。
      schema: {type: string, minLength: 8, maxLength: 128}
    Page: {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
    Size: {in: query, name: size, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  schemas:
    Envelope:
      type: object
      required: [code, message, data, requestId]
      properties:
        code: {$ref: '#/components/schemas/ApiCode'}
        message: {type: string, example: ok}
        data: {}
        requestId: {type: string}
    ApiCode:
      description: |
        业务结果码，所有响应的 code 字段取值。HTTP 状态码与之一一对应：
        OK=200；VALIDATION_FAILED=400；UNAUTHENTICATED=401；FORBIDDEN/HUMAN_ONLY_PERIOD=403；
        NOT_FOUND=404；METHOD_NOT_ALLOWED=405；其余冲突类=409；RATE_LIMITED=429；INTERNAL_ERROR=500。
        客户端应以 code 判定业务结果，不要只依赖 HTTP 状态码。
        两个成功码 OK 与 IDEMPOTENCY_REPLAY 都对应 200。
      type: string
      enum:
        - OK
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - INVALID_STATE
        - BID_TOO_LOW
        - BID_LATE
        - HUMAN_ONLY_PERIOD
        - NOT_JOINED
        - INSUFFICIENT_BALANCE
        - IDEMPOTENCY_REPLAY
        - CONFLICT
        - METHOD_NOT_ALLOWED
        - RATE_LIMITED
        - INTERNAL_ERROR
...
    User:
      type: object
      required: [id, name, role]
      properties:
        id: {type: string}
        name: {type: string}
        role: {type: string, enum: [BIDDER, ADMIN]}
    AuctionStatus: {type: string, enum: [DRAFT, RUNNING, SETTLING, FINISHED, CANCELLED]}
    AuctionSnapshot:
      type: object
      required: [id, title, status, startPrice, minIncrement, currentPrice, extensionCount, participantCount, seq, finalGameWindowSeconds, serverTime]
      properties:
        id: {type: string}
        title: {type: string}
        description: {type: string}
        status: {$ref: '#/components/schemas/AuctionStatus'}
        startPrice: {type: integer, minimum: 0}
        minIncrement: {type: integer, minimum: 1}
        currentPrice: {type: integer, minimum: 0}
        leader: {type: string, nullable: true}
        endsAt: {type: string, format: date-time, nullable: true}
        extensionCount: {type: integer, minimum: 0}
        participantCount: {type: integer, minimum: 0}
        seq: {type: integer, minimum: 0}
        finalGameWindowSeconds:
          type: integer
          minimum: 0
          description: |
            尾段“博弈时间”长度（秒）：截止前这段时间内 Agent 一律不得出价，只有真人可以。
            由服务端给出而不是让前端硬编码：窗口长度是运营配置（`AUCTION_FINAL_GAME_WINDOW_SECONDS`），
            前端只用它做“进入博弈时间”的提示，**不承担任何判定职责**——真正的拒绝对话发生在服务端事务内。
        serverTime: {type: string, format: date-time}
    BidRequest:
      type: object
      required: [requestId, amount]
      properties:
        requestId: {type: string, minLength: 8, maxLength: 128}
        amount: {type: integer, minimum: 1}
    BidResult:
      type: object
      required: [accepted, idempotent, price, extensions, seq, serverTime]
      properties:
        accepted: {type: boolean}
        idempotent: {type: boolean}
        price: {type: integer}
        leader: {type: string, nullable: true}
        extensions: {type: integer}
        seq: {type: integer}
        serverTime: {type: string, format: date-time}
    Bid:
      description: 一条已接受的出价记录。seq 单调递增，客户端据此检测事件缺口。
      type: object
      required: [id, userId, amount, requestId, seq, serverTime]
      properties:
        id: {type: string}
        userId: {type: string}
        amount: {type: integer, minimum: 1}
        requestId: {type: string}
        seq: {type: integer, minimum: 1}
        serverTime: {type: string, format: date-time}
    AuctionPage:
      type: object
      required: [items, page, size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/AuctionSnapshot'}}
        page: {type: integer, minimum: 1}
        size: {type: integer, minimum: 1}
        total: {type: integer, minimum: 0}
    BidPage:
      type: object
      required: [items, page, size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Bid'}}
        page: {type: integer, minimum: 1}
        size: {type: integer, minimum: 1}
        total: {type: integer, minimum: 0}
...
    CreateAuctionRequest:
      type: object
      required: [title, startPrice, minIncrement, durationSeconds]
      properties:
        title: {type: string, minLength: 1, maxLength: 120}
        description: {type: string, maxLength: 2000}
        startPrice: {type: integer, minimum: 1}
        minIncrement: {type: integer, minimum: 1}
        durationSeconds: {type: integer, minimum: 10, maximum: 86400}
...
    Participant: {type: object, properties: {auctionId: {type: string}, userId: {type: string}, joinedAt: {type: string, format: date-time}}}
    AuctionResult:
      type: object
      required: [auctionId, status, finalPrice, reason]
      properties:
        auctionId: {type: string}
        status: {$ref: '#/components/schemas/AuctionStatus'}
        winner: {type: string, nullable: true}
        winnerType:
          type: string
          enum: [HUMAN, AGENT]
          nullable: true
          description: |
            成交主体：HUMAN=真人，AGENT=竞拍 Agent；未成交时为 null。
            这是隐私字段——仅当请求者是赢家本人或管理员时才有值，其余一律为 null。
        finalPrice: {type: integer, minimum: 0}
        reason: {type: string, enum: [TIMEOUT, NO_BIDS, CANCELLED]}
        settledAt: {type: string, format: date-time, nullable: true}
...
  responses:
...
    EnvelopeUser: {description: OK, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/User'}}}]}}}}
...
    EnvelopeAuction: {description: OK, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/AuctionSnapshot'}}}]}}}}
    EnvelopeAuctionPage: {description: OK, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/AuctionPage'}}}]}}}}
    EnvelopeBidPage: {description: OK, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/BidPage'}}}]}}}}
    EnvelopeBidResult: {description: OK, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/BidResult'}}}]}}}}
    EnvelopeResult: {description: OK, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/AuctionResult'}}}]}}}}
    EnvelopeParticipant: {description: OK, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/Participant'}}}]}}}}
...
    BadRequest: {description: Invalid request, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    Unauthorized: {description: Authentication required, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    Forbidden: {description: Permission denied, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    Conflict: {description: Business conflict such as BID_TOO_LOW, BID_LATE, NOT_JOINED, INVALID_STATE, or IDEMPOTENCY_REPLAY, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    RateLimited: {description: Rate limited, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}